+++ src/@lib/equalities/shallowEquals.ts
@@ -1,17 +1,5 @@
-import { isObject } from "./utils";
+import { baseEquals } from "./utils";
 
 export function shallowEquals<T>(objA: T, objB: T): boolean {
-  if (Array.isArray(objA) && Array.isArray(objB)) {
-    if (objA.length !== objB.length) return false;
-    return objA.every((v, idx) => v === objB[idx]);
-  }
-
-  if (isObject(objA) && isObject(objB)) {
-    const aKeys = Object.keys(objA);
-    const bKeys = Object.keys(objB);
-    if (aKeys.length !== bKeys.length) return false;
-    return aKeys.every((key) => objA[key] === objB[key]);
-  }
-
-  return objA === objB;
+  return baseEquals(objA, objB, (a, b) => a === b);
 }
