+++ src/@lib/hooks/useCallback.ts
@@ -4,7 +4,7 @@ import { useMemo } from "./useMemo";
 
 export function useCallback<T extends Function>(
   factory: T,
-  _deps: DependencyList,
+  _deps: DependencyList = [],
 ) {
   return useMemo(() => factory, _deps);
 }
